+++ es6/expressions.d.ts
@@ -15,10 +15,16 @@ interface ParserOptions {
   ) => any;
 }
 
+interface ExpressionParser extends DXT.Parser {
+  compiled: any;
+  getIdentifiers(): string[];
+  getObjectIdentifiers(): any;
+}
+
 type Parser = {
-  (tag: string): DXT.Parser;
+  (tag: string): ExpressionParser;
   filters: { [x: string]: (input: any, ...filters: any[]) => any };
-  configure: (options: ParserOptions) => (tag: string) => DXT.Parser;
+  configure: (options: ParserOptions) => (tag: string) => ExpressionParser;
 };
 
 declare var expressionParser: Parser;